Move rate limiting to separate module
